♻️ refactor: edit validation nomenclature
